Cleanup. Move filesystem and listener funcs to utils.go
