+++ main.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"flag"
 	"fmt"
-	"github.com/BurntSushi/toml"
 	"github.com/NYTimes/gziphandler"
 	"github.com/coreos/go-systemd/activation"
 	"github.com/coreos/go-systemd/daemon"
@@ -16,9 +15,7 @@ import (
 	"github.com/mpdroog/ratelimit"
 	"github.com/mpdroog/ratelimit/memory"
 	"golang.org/x/crypto/acme/autocert"
-	"golang.org/x/net/netutil"
 	"golang.org/x/text/language"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"net/http/pprof"
@@ -29,60 +26,6 @@ import (
 	"syscall"
 	"time"
 )
-
-func getDomains() ([]string, error) {
-	fileInfo, err := ioutil.ReadDir(config.Webdir)
-	if err != nil {
-		return nil, err
-	}
-
-	out := []string{}
-	for _, file := range fileInfo {
-		if file.IsDir() {
-			if strings.ToLower(file.Name()) != file.Name() {
-				return nil, fmt.Errorf(config.Webdir + file.Name() + " not lowercase!")
-			}
-			out = append(out, file.Name())
-		}
-	}
-	return out, nil
-}
-
-func exists(path string) (bool, error) {
-	_, err := os.Stat(path)
-	if err == nil {
-		return true, nil
-	}
-	if os.IsNotExist(err) {
-		return false, nil
-	}
-	return true, err
-}
-
-func getOverride(path string) (config.Override, error) {
-	c := config.Override{}
-
-	if _, e := os.Stat(path); os.IsNotExist(e) {
-		return c, nil
-	}
-
-	r, e := os.Open(path)
-	if e != nil {
-		return c, e
-	}
-	defer r.Close()
-	_, e = toml.DecodeReader(r, &c)
-	return c, e
-}
-
-func listener(addr string) (net.Listener, error) {
-	ln, e := net.Listen("tcp", addr)
-	if e != nil {
-		return nil, e
-	}
-	aln := TCPKeepAliveListener{ln.(*net.TCPListener)}
-	return netutil.LimitListener(aln, config.MAX_WORKERS), nil
-}
 
 func main() {
 	skipsysd := false
